Replaced landerCompass with landerDOF and dofCompass to prepare for using the other 10 DOF sensors, such as the altimeter.
